backend/requirements.txt
fastapi
uvicorn
transformers
torch
wikipedia
pymupdf
pydantic
python-dotenv
requests
cohere
google-generativeai
typing-extensions
python-multipart